Fix parsing various things out of pom.xml
Ox.locate returns an array so we need to look at the first element
of the array for a lot of things. Also take advantage of the fact
that Ox.locate can take a path in the xml instead of constructing
that logic.

Add tests with a sample pom file to show things are working