add endpoints to efficiently get only project deps / bulk project deps
/:platform/:name/:version/dependencies?subset=minimum endpoint gets
only the deps for project without all the other project data.

/projects/dependencies endpoint is a bulk version with the same subset
parameter available.

Had to factor the code from application_controller for looking up projects
out into a method on Project so it could be used for the bulk endpoint.

Also avoided using the application_controller.find_project for the
non-bulk /dependencies endpoint since the subset parameter changes
how we want to do the query (whether we need to include versions).