Included updated_at in package API replies and add /updated endpoint
The /updated endpoint returns a list of packages with updated_at
in a given time window.
The purpose of this is to simplify mirroring.

Only add experimental updated_at field if using an internal API key
We don't want anyone to rely on this field yet until we see how it works out.

Track a deleted_projects table with ONLY the project name and time of deletion
All other information gets discarded.

Keep only a digest of deleted project names, not recoverable name