Add package-deprecation support for pypi yanking.
Pypi is similar to NPM in that it only allows specific releases to
be "yanked", but having the latest release yanked is treated
as if the whole package is yanked.

Also, according to PEP 592, yanked packages are still downloadable,
and they won't be used by package managers unless they're the only
possible release that satisfies a manifest.