Fix incorrectly stored Maven dependency platform information
Dependency platform names for Maven dependencies were being incorrectly
stored using the name of the repository from which the original project
was being found within, and not the name of the platform itself. This PR
ensures no new dependencies are added with incorrect platforms, and
provides a Rake task that can be run in batches to repair existing Maven
dependencies.