Improve NuGet data retrieval, fix deprecation bug
This implements for NuGet the same class structure approach we used in PyPI.
This made it easier to find and fix, and write tests for, an issue with
package deprecation where the incorrect release was being used. As with
the PyPI changes, this removes more uses of raw Hashes for data storage
and transfer around the app.