Handle weird case where latest in maven-metadata doesn't exist
There are some cases (for example, commons-discovery:commons-discovery)
where the version listed as latest in maven-metadata.xml isn't actually
present on central. This lets us fallback to our scraping and using
that.

Also improves the scraping of the latest version to prefer non
date formatted versions as that's an older maven practice