ProjectUpdatedSerializer: send fields that are needed to search for the Project
We want to send only "cheap" fields (in particular: avoid joining to Repository
or Version), but include fields that are useful to search against in order to
discover relevant projects.

We throw in the release numbers just so we have a sample release of the
project if we need it.

Some of these are in here in case they are useful to sort search results.