ProjectUpdatedWorker: don't ignore errors on all_project_updates
We had set ignore_errors:true to reduce logging noise, but unfortunately
it means this job never retries and enver exhausts retries. To retry or
to exhaust retries we need to raise an exception.

While there have been no errors from the all_projet_updates hooks that
I see in the past 24 hours, we do want to be sure we retry if there
ever are such errors.